gains component created and implemented in the exercises page, still doesn't work for multiple components
