+++ src/components/gains-chart/gains-chart.html
@@ -1,6 +1,6 @@
-<div>
-	<ion-item no-lines>
+<!-- Generated template for the GainsChartComponent component -->
+<ion-item no-lines>
+	<div id="chart">
 		<svg id="gainsChart"></svg>
-	</ion-item>
-	
-</div>
+	</div>
+</ion-item>
