+++ src/pages/exercises/exercises.module.ts
@@ -1,13 +1,14 @@
 import { NgModule } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
 import { IonicPageModule } from 'ionic-angular';
-
+import { GainsChartComponent } from '../../components/gains-chart/gains-chart';
 import { ListMasterPage } from './exercises';
 import { PipesModule } from '../../pipes/pipes.module';
 
 @NgModule({
   declarations: [
     ListMasterPage,
+    GainsChartComponent
   ],
   imports: [
     IonicPageModule.forChild(ListMasterPage),
