+++ src/pages/gains/gains.html
@@ -14,9 +14,5 @@
 
 <ion-content padding>
 	
-	<ion-item no-lines>
-		<div id="chart">
-			<svg id="gainsChart"></svg>
-		</div>
-	</ion-item>
+	<gains-chart></gains-chart>
 </ion-content>
