+++ src/pages/profile/profile.ts
@@ -181,7 +181,7 @@ export class SettingsPage {
       this.form.patchValue({"profilePic": "data:image/jpeg;base64," + data});
       if (data != "NahNigga"){
         this.show = true;
-        console.log(data);
+        //console.log(data);
         this.form.patchValue({"profilePic": "data:image/jpeg;base64," + data});
       } else {
         // <ion-icon *ngIf="item.profilePic == 'data:image/jpeg;base64,NahNigga'" class="default-img" name="contact"></ion-icon>
